backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
openai==1.6.1
PyJWT==2.8.0
cryptography==41.0.7
pandas==2.1.4
python-docx==1.1.0
PyPDF2==3.0.1
requests==2.31.0
gunicorn==21.2.0

